+++ philo/philosopher.h
@@ -31,7 +31,7 @@ typedef struct s_philo {
 	long				idx;
 	enum e_philo_status	status;
 	long				last_eating_ms;
-	int					eating_count;
+	long				eating_count;
 	bool				is_living;
 	pthread_mutex_t		mux;
 	pthread_t			thread;
@@ -47,18 +47,18 @@ bool	is_philo_still_alive(t_philo *philo);
 bool	has_philo_eaten_n_times(t_philo *philo);
 int		philosopher_eat(t_philos_info *philos_info, t_philo *philo);
 int		philosopher_sleep(t_philos_info *philos_info, t_philo *philo);
-void	write_philo_status(int philo_idx, enum e_philo_status status, long ms);
+void	write_philo_status(long philo_idx, enum e_philo_status status, long ms);
 // philosopher observer
 int		start_philo_observers(t_philos_info *philos_info,
 			t_philo *philos, pthread_t *philo_observers);
 int		wait_philo_observers(t_philos_info *philos_info,
 			pthread_t *philo_observers);
 // fork
-int		init_forks(pthread_mutex_t *forks, int fork_num);
-void	hold_fork(pthread_mutex_t *forks, int fork_idx);
-void	release_fork(pthread_mutex_t *forks, int fork_idx);
-void	release_forks(pthread_mutex_t *forks, int fork_idx1, int fork_idx2);
+int		init_forks(pthread_mutex_t *forks, long fork_num);
+void	hold_fork(pthread_mutex_t *forks, long fork_idx);
+void	release_fork(pthread_mutex_t *forks, long fork_idx);
+void	release_forks(pthread_mutex_t *forks, long fork_idx1, long fork_idx2);
 int		release_forks_and_rtn_err(pthread_mutex_t *forks,
-			int fork_idx1, int fork_idx2);
+			long fork_idx1, long fork_idx2);
 
 #endif
